requirements.txt
Pillow
pygame
yt-dlp
python-vlc
imageio-ffmpeg